page for creating a new user
