+++ cmd/server/main.go
@@ -161,8 +161,9 @@ func realMain(ctx context.Context) error {
 			sub.Use(rateLimit)
 
 			sub.Handle("/", loginController.HandleLogin()).Methods("GET")
+			sub.Handle("/login/create", loginController.HandleLoginCreate()).Methods("GET")
 			sub.Handle("/legacylogin", loginController.HandleLegacyLogin()).Methods("GET")
-			sub.Handle("/session", loginController.HandleCreate()).Methods("POST")
+			sub.Handle("/session", loginController.HandleCreateSession()).Methods("POST")
 			sub.Handle("/signout", loginController.HandleSignOut()).Methods("GET")
 		}
