+++ pkg/controller/login/create.go
@@ -12,50 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package login defines the controller for the login page.
 package login
 
 import (
 	"net/http"
 
-	"github.com/google/exposure-notifications-verification-server/pkg/api"
 	"github.com/google/exposure-notifications-verification-server/pkg/controller"
 )
 
-func (c *Controller) HandleCreate() http.Handler {
-	type FormData struct {
-		IDToken string `form:"idToken,required"`
-	}
-
+func (c *Controller) HandleLoginCreate() http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		ctx := r.Context()
 
+		// If there's a firebase cookie in the session, try to redirect to /home. If
+		// the cookie is invalid, the auth middleware will pick it up, delete the
+		// cookie from the session, and kick them back here.
 		session := controller.SessionFromContext(ctx)
-		if session == nil {
-			controller.MissingSession(w, r, c.h)
-			return
-		}
-		flash := controller.Flash(session)
-
-		// Parse and decode form.
-		var form FormData
-		if err := controller.BindForm(w, r, &form); err != nil {
-			flash.Error("Failed to process form: %v", err)
-			c.h.RenderJSON(w, http.StatusBadRequest, api.Error(err))
-			return
+		if session != nil {
+			if c := controller.FirebaseCookieFromSession(session); c != "" {
+				http.Redirect(w, r, "/home", http.StatusSeeOther)
+				return
+			}
 		}
 
-		// Get the session cookie from firebase.
-		ttl := c.config.SessionDuration
-		cookie, err := c.client.SessionCookie(ctx, form.IDToken, ttl)
-		if err != nil {
-			flash.Error("Failed to create session: %v", err)
-			c.h.RenderJSON(w, http.StatusUnauthorized, api.Error(err))
-			return
-		}
-
-		// Set the firebase cookie value in our session.
-		controller.StoreSessionFirebaseCookie(session, cookie)
-
-		c.h.RenderJSON(w, http.StatusOK, nil)
+		m := controller.TemplateMapFromContext(ctx)
+		m["firebase"] = c.config.Firebase
+		c.h.RenderHTML(w, "login/create", m)
 	})
 }
